Makes Pokeblock free
Goal: Reduce the grind of hub metaprogression.

This eliminates the button-mashing of Berry farming, gives infinite money (in a roundabout way), and allows players to increase their partners' IVs arbitrarily.

Again, my knowledge of C and of the Rogue code base is limited, so this is a hacky fix. I just want to play a full run in one session.